Fix adding LoadModule statement when missing
The script is relying on existence of "native osx" module as reference
point to add a missing PHP module, but when the line does not exist in
the httpd.conf file (which is the case with macOS 13), then the
LoadModule line is never added resulting in a non-functional web server
that can't handle PHP scripts.

Using rewrite_module as reference point instead.

Fixes #6